refactor: remove repeating types
