+++ keep-ui/entities/workflows/model/useWorkflowActions.ts
@@ -2,7 +2,7 @@ import { useApi } from "@/shared/lib/hooks/useApi";
 import { showSuccessToast } from "@/shared/ui/utils/showSuccessToast";
 import { useRevalidateMultiple } from "@/shared/lib/state-utils";
 import { showErrorToast } from "@/shared/ui";
-import { Definition } from "@/app/(keep)/workflows/builder/builder-store";
+import { Definition } from "@/app/(keep)/workflows/builder/types";
 import { getWorkflowFromDefinition } from "@/app/(keep)/workflows/builder/utils";
 import { stringify } from "yaml";
 import { useCallback } from "react";
